Thêm phân quyền chỉ có chủ nhóm mới được phép xóa sửa tên thành viên và xóa hóa đơn

+++ Project_BeFair/src/main/java/khanh/ntu/BF/controllers/ExpenseController.java
@@ -1,5 +1,6 @@
 package khanh.ntu.BF.controllers;
 
+import java.security.Principal;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import khanh.ntu.BF.services.BeFairService;
 
@@ -32,7 +34,13 @@ public class ExpenseController {
     
     //Xóa hóa đơn
     @PostMapping("/group/{groupId}/delete-expense/{expenseId}")
-    public String deleteExpense(@PathVariable Long groupId, @PathVariable Long expenseId) {
+    public String deleteExpense(@PathVariable Long groupId, @PathVariable Long expenseId, Principal principal, RedirectAttributes redirectAttributes) {
+    	
+    	if (!bfService.isGroupOwner(groupId, principal.getName())) {
+            redirectAttributes.addFlashAttribute("errorMessage", "Bạn không phải chủ nhóm, không có quyền xóa hóa đơn!");
+            return "redirect:/group/" + groupId;
+        }
+    	
         bfService.deleteExpense(expenseId);
         return "redirect:/group/" + groupId;
     }
